# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCluster Control Plane API
  version: 0.1.0
  description: HTTP contract for the OpenCluster product API and inbound webhook endpoints.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local self-hosted deployment
...
paths:
...
  /api/v1/organizations:
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: authenticated
      x-opencluster-organization-selector: forbidden
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      responses:
        '200':
          description: Organizations available to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: listOrganizations
      summary: List organizations
      tags:
        - Organizations
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
    post:
      security:
        - SessionCookie: []
      x-opencluster-access: authenticated
      x-opencluster-organization-selector: forbidden
      x-opencluster-cookie-csrf: required-for-unsafe-cookie-request
      x-opencluster-idempotency: not-supported
      responses:
        '201':
          description: Organization and initial Admin membership created atomically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      operationId: createOrganization
      summary: Create organization
      tags:
        - Organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
        required: true
...
  /api/v1/relays:
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/RelaySearch'
        - $ref: '#/components/parameters/RelaySort'
        - $ref: '#/components/parameters/RelayStateFilter'
        - $ref: '#/components/parameters/RelayVersionFilter'
        - $ref: '#/components/parameters/RelayCapabilityFilter'
      responses:
        '200':
          $ref: '#/components/responses/RelayList'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: listRelays
      summary: List relays
      tags:
        - Relays
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-oc_session
    BootstrapToken:
      type: http
      scheme: bearer
      description: Deployment bootstrap credential; accepted only until bootstrap succeeds.
    WebhookToken:
      type: apiKey
      in: header
      name: X-OpenCluster-Token
      description: Static sender credential. It authenticates the sender and does not sign the body.
    SlackSignature:
      type: apiKey
      in: header
      name: X-Slack-Signature
      description: Slack request signature verified with the configured signing secret.
  parameters:
    Organization:
      name: X-OpenCluster-Organization
      in: header
      required: true
      description: Selects the Organization whose data and membership are authorized.
      schema:
        $ref: '#/components/schemas/OrganizationID'
      example: 7d722f19-1978-4dc1-a176-e9be92b7e154
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque resume position returned in the preceding page's next field.
      schema:
        type: string
        minLength: 1
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    RelaySearch:
      name: search
      in: query
      required: false
      description: Case-insensitive containment search over the Relay identifier, fingerprint, and version.
      schema:
        type: string
        minLength: 1
        maxLength: 256
...
    RelaySort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum:
          - registeredAt
          - -registeredAt
          - lastSeenAt
          - -lastSeenAt
          - version
          - -version
          - fingerprint
          - -fingerprint
...
    RelayStateFilter:
      name: state
      in: query
      required: false
      schema:
        type: string
        enum:
          - connected
          - disconnected
          - revoked
          - degraded
    RelayVersionFilter:
      name: version
      in: query
      required: false
      description: Exact Relay version without surrounding whitespace.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^\S(?:.*\S)?$
    RelayCapabilityFilter:
      name: capability
      in: query
      required: false
      description: Exact capability identifier without surrounding whitespace.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^\S(?:.*\S)?$
...
  responses:
    ErrorResponse:
      description: The operation did not complete.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RelayList:
      description: Relay roster.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RelayList'
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or rejected credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Membership, permission, Organization selection, or cookie origin was rejected.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: The requested state conflicts with durable state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    OrganizationID:
      type: string
      format: uuid
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 1024
          example: the selected Organization is not available to this caller
        reason:
          type: string
          enum:
            - credential_rejected
            - session_expired
            - session_revoked
        requires:
          type: string
          minLength: 1
          maxLength: 128
...
    Organization:
      type: object
      additionalProperties: false
      required:
        - id
        - displayName
        - membership
      properties:
        id:
          $ref: '#/components/schemas/OrganizationID'
        displayName:
          type: string
          minLength: 1
          maxLength: 256
        membership:
          $ref: '#/components/schemas/OrganizationMembership'
    OrganizationMembership:
      type: object
      additionalProperties: false
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - admin
            - editor
            - viewer
...
    OrganizationList:
      type: object
      additionalProperties: false
      required:
        - organizations
        - next
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        next:
          oneOf:
            - type: string
              maxLength: 512
            - type: 'null'
    CreateOrganizationRequest:
      type: object
      additionalProperties: false
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 256
...
    Relay:
      type: object
      additionalProperties: false
      required:
        - registrationId
        - clusterFingerprint
        - relayVersion
        - compatibility
        - registeredAt
        - connected
        - capabilities
      properties:
        registrationId:
          $ref: '#/components/schemas/UUID'
        clusterFingerprint:
          type: string
          minLength: 1
          maxLength: 256
        relayVersion:
          type: string
          minLength: 1
          maxLength: 64
        compatibility:
          type: string
          enum:
            - compatible
            - unknown
        registeredAt:
          $ref: '#/components/schemas/Timestamp'
        revokedAt:
          $ref: '#/components/schemas/Timestamp'
        connected:
          type: boolean
        lastSeenAt:
          $ref: '#/components/schemas/Timestamp'
        capabilities:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
          uniqueItems: true
        sessionConflict:
          type: object
          additionalProperties: false
          required:
            - detectedAt
            - distinctHosts
            - multipleHosts
          properties:
            detectedAt:
              $ref: '#/components/schemas/Timestamp'
            distinctHosts:
              type: integer
              minimum: 2
            multipleHosts:
              type: boolean
              const: true
    RelayList:
      type: object
      additionalProperties: false
      required:
        - items
        - next
        - total
      properties:
        next:
          oneOf:
            - type: string
              maxLength: 512
            - type: 'null'
        total:
          oneOf:
            - type: integer
              minimum: 0
            - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Relay'
...